requirements.txt
pymupdf
chromadb
keybert
sentence-transformers
fpdf
prettytable
langchain
langchain_experimental
pandas
langgraph
langchain-google-genai
google-generativeai
reportlab
python-dotenv
pywhatkit
openpyxl
twilio
smtplib
email